requirements.txt
openai==0.28
chainlit
pandas
python-dotenv
flask